moved flatfiletemp to flatfile to make it main web interface
